+++ backend/src/main/java/com/woowacourse/momo/domain/group/Duration.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.domain;
+package com.woowacourse.momo.domain.group;
 
 import lombok.NoArgsConstructor;
