+++ backend/src/main/java/com/woowacourse/momo/domain/group/Schedule.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.domain;
+package com.woowacourse.momo.domain.group;
 
 import lombok.Getter;
