+++ backend/src/main/java/com/woowacourse/momo/domain/member/Member.java
@@ -1,12 +1,14 @@
-package com.woowacourse.momo.domain;
+package com.woowacourse.momo.domain.member;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
 @Getter
+@NoArgsConstructor
 @Entity
-public class Category {
+public class Member {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -14,4 +16,8 @@ public class Category {
 
     @Column(length = 30, nullable = false)
     private String name;
+
+    public Member(String name) {
+        this.name = name;
+    }
 }
