+++ backend/src/test/java/com/woowacourse/momo/domain/group/DurationTest.java
@@ -1,5 +1,6 @@
-package com.woowacourse.momo.domain;
+package com.woowacourse.momo.domain.group;
 
+import com.woowacourse.momo.domain.group.Duration;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
