+++ backend/src/test/java/com/woowacourse/momo/domain/group/GroupTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.domain;
+package com.woowacourse.momo.domain.group;
 
 import static org.junit.jupiter.api.Assertions.*;
